+++ SCM.Application/Models/DTOs/Departments/DepartmentDTO.cs
@@ -4,9 +4,9 @@ namespace SCM.Application.Models.DTOs.Departments
 {
     public class DepartmentDTO
     {
-        public BigInteger Id { get; set; }
+        public Int64 Id { get; set; }
         public string Name { get; set; }
-        public BigInteger CompanyId { get; set; }
+        public Int64 CompanyId { get; set; }
         public string CompanyName { get; set; }
     }
 }
