+++ SCM.Application/Models/DTOs/Requests/RequestDTO.cs
@@ -5,8 +5,8 @@ namespace SCM.Application.Models.DTOs.Requests
 {
     public class RequestDTO
     {
-        public BigInteger Id { get; set; }
-        public BigInteger UserId { get; set; }
+        public Int64 Id { get; set; }
+        public Int64 UserId { get; set; }
         public string UserName { get; set; }
         public DateTime RequestDate { get; set; }
         public RequestStatus Status { get; set; }
