+++ SCM.Application/Models/DTOs/Suppliers/SupplierDTO.cs
@@ -4,11 +4,11 @@ namespace SCM.Application.Models.DTOs.Suppliers
 {
     public class SupplierDTO
     {
-        public BigInteger Id { get; set; }
+        public Int64 Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
         public string Phone { get; set; }
-        public BigInteger RoleId { get; set; }
+        public Int64 RoleId { get; set; }
         public string RoleName { get; set; }
     }
 }
