+++ SCM.Application/Models/RequestModels/Approves/ApproveVM.cs
@@ -4,7 +4,7 @@ namespace SCM.Application.Models.RequestModels.Approves
 {
     public class ApproveVM
     {
-        public BigInteger RequestId { get; set; }
+        public Int64 RequestId { get; set; }
         public bool IsApproved { get; set; }
     }
 }
