+++ SCM.Application/Models/RequestModels/Companies/UpdateCompanyVM.cs
@@ -4,7 +4,7 @@ namespace SCM.Application.Models.RequestModels.Companies
 {
     public class UpdateCompanyVM
     {
-        public BigInteger Id { get; set; }
+        public Int64 Id { get; set; }
         public string CompanyName { get; set; }
     }
 }
