+++ SCM.Application/Models/RequestModels/Invoice/GetInvoiceVM.cs
@@ -4,6 +4,6 @@ namespace SCM.Application.Models.RequestModels.Invoice
 {
     public class GetInvoiceVM
     {
-        public BigInteger Id { get; set; }
+        public Int64 Id { get; set; }
     }
 }
