+++ SCM.Application/Models/RequestModels/Products/UpdateProductVM.cs
@@ -4,11 +4,11 @@ namespace SCM.Application.Models.RequestModels.Products
 {
     public class UpdateProductVM
     {
-        public BigInteger Id { get; set; }
-        public BigInteger? CategoryId { get; set; }
+        public Int64 Id { get; set; }
+        public Int64? CategoryId { get; set; }
         public string Name { get; set; }
         public string Detail { get; set; }
-        public BigInteger? UnitInStock { get; set; }
+        public Int64? UnitInStock { get; set; }
         public decimal? UnitPrice { get; set; }
     }
 }
