+++ SCM.Application/Services/Implementations/ProductService.cs
@@ -10,7 +10,6 @@ using SCM.Application.Validators.Products;
 using SCM.Application.Wrapper;
 using SCM.Domain.Entities;
 using SCM.Domain.UnitofWork;
-using System.Numerics;
 
 namespace SCM.Application.Services.Implementations
 {
@@ -52,9 +51,9 @@ namespace SCM.Application.Services.Implementations
 
         #region Insert, Update, Delete
         [ValidationBehavior(typeof(CreateProductValidator))]
-        public async Task<Result<BigInteger>> CreateProduct(CreateProductVM createProductVM)
+        public async Task<Result<Int64>> CreateProduct(CreateProductVM createProductVM)
         {
-            var result = new Result<BigInteger>();
+            var result = new Result<Int64>();
 
             var productExistsSameName = await _uWork.GetRepository<Product>().AnyAsync(x => x.Name == createProductVM.Name.Trim());
             if (productExistsSameName)
@@ -71,9 +70,9 @@ namespace SCM.Application.Services.Implementations
         }
 
         [ValidationBehavior(typeof(DeleteProductValidator))]
-        public async Task<Result<BigInteger>> DeleteProduct(DeleteProductVM deleteProductVM)
+        public async Task<Result<Int64>> DeleteProduct(DeleteProductVM deleteProductVM)
         {
-            var result = new Result<BigInteger>();
+            var result = new Result<Int64>();
 
             var productEntity = await _uWork.GetRepository<Product>().GetById(deleteProductVM.Id);
             if (productEntity is null)
@@ -89,9 +88,9 @@ namespace SCM.Application.Services.Implementations
         }
 
         [ValidationBehavior(typeof(UpdateProductValidator))]
-        public async Task<Result<BigInteger>> UpdateProduct(UpdateProductVM updateProductVM)
+        public async Task<Result<Int64>> UpdateProduct(UpdateProductVM updateProductVM)
         {
-            var result = new Result<BigInteger>();
+            var result = new Result<Int64>();
 
             var productEntity = await _uWork.GetRepository<Product>().GetById(updateProductVM.Id);
             if (productEntity is null)
