+++ SCM.Domain/Entities/Approves.cs
@@ -10,9 +10,9 @@ namespace SCM.Domain.Entities
             Requests=new List<Request>();
         }
 
-        public BigInteger RequestId { get; set; }
+        public Int64 RequestId { get; set; }
         public decimal Amount { get; set; }
-        public BigInteger ApproveId { get; set; }
+        public Int64 ApproveId { get; set; }
         public RequestStatus Status { get; set; }
         public IEnumerable<Request> Requests { get; set; }
     }
