+++ SCM.Domain/Entities/Employee.cs
@@ -11,9 +11,9 @@ namespace SCM.Domain.Entities
         public string Email { get; set; }
         public string Phone { get; set; }
         public DateTime BirthDate { get; set; }
-        public BigInteger CompanyId { get; set; }
-        public BigInteger DepartmentId { get; set; }
-
+        public Int64 CompanyId { get; set; }
+        public Int64 DepartmentId { get; set; }
+            
 
         public virtual Account Account { get; set; }
         public virtual ICollection<Request> Requests { get; set; }
