+++ SCM.Domain/Entities/Message.cs
@@ -5,7 +5,7 @@ namespace SCM.Domain.Entities
 {
     public class Message : AuditableEntity
     {
-        public BigInteger UserId { get; set; }
+        public Int64 UserId { get; set; }
         public string Description { get; set; }
         public Company Company { get; set; }
         public Department Department { get; set; }
